+++ waterfall-cad-examples/src/OffsetExample.hs
@@ -12,7 +12,7 @@ offsetExample =
   let beam axis = rotate axis (pi/2) (scale (V3 2 2 4) Solids.centeredCube)
       cross = foldMap beam [unit _x,  unit _y, unit _z]
       tolerance = 1e-6
-      offsetCross amount = offset amount tolerance cross
+      offsetCross amount = offset tolerance amount cross
       offsetCrosses = offsetCross <$> [- 0.5, - 0.25, 0, 0.25, 0.5] 
       position = (unit _x ^* 5 ^*) <$> [0..]  
   in mconcat $ zipWith translate position offsetCrosses